issues with closing file in FROM at a high QOS
For rapid writing to a pipe, for example logging sensor data to a pipe, and periodically dumping the pipe to a file, then continuing to write to the pipe without loss of data is difficult without pausing motion tasks. An extra flag is added to files__write_from_pipe to designate if the function should handle the opening and closing of the file, or if the user will manually do it outside of the scope of the function. Opening/closing the file outside of writing function should help with speed. However if this function is used over multiple programs for the same file it will not work.